Split analyze_inst into instruction-category methods
Refactor the massive 2,400-line analyze_inst method in rue-air/src/sema/analysis.rs
into 8-10 category methods for better maintainability.

Changes:
- Create new analyze_ops.rs module with category dispatch methods:
  - analyze_literal (IntConst, BoolConst, StringConst, UnitConst)
  - analyze_unary_op (Neg, Not, BitNot)
  - analyze_logical_op (And, Or)
  - analyze_control_flow (If, Loop, While, Block, etc.)
  - analyze_variable_ops (VarRef, ParamRef, VarDecl, etc.)
  - analyze_struct_ops (StructInit, FieldGet, FieldSet)
  - analyze_array_ops (ArrayInit, IndexGet, IndexSet)
  - analyze_enum_ops (EnumVariant, Match)
  - analyze_call_ops (Call, MethodCall, AssocFnCall)
  - analyze_intrinsic_ops (Intrinsic, TypeIntrinsic)
  - analyze_decl_noop (ImplDecl, DropFnDecl, FnDecl)

- analyze_inst is now ~130 lines (pure dispatch logic)
- Each category method under 500 lines
- Complex operations delegate to _impl methods in analysis.rs
- All existing tests pass
- No functionality changes (pure refactor)

Closes: rue-me8s
